Monstre + MonstreService

+++ SufodBack/src/main/java/sufod/entity/Monstre.java
@@ -7,33 +7,30 @@ import javax.persistence.OneToMany;
 import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
 
-
 @Entity
-@Table(name="monstre")
+@Table(name = "monstre")
 @SequenceGenerator(name = "seqMonstre", sequenceName = "seq_monstre", initialValue = 10, allocationSize = 1)
-public class Monstre extends Vivant{
-	
-	//@Id
-	//@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seqMonstre")
+public class Monstre extends Vivant {
+
+	// @Id
+	// @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seqMonstre")
 	/*----------- Attributs -----------*/
-	@OneToMany(mappedBy="monstre")
+	@OneToMany(mappedBy = "monstre")
 	private List<Attaque> attaque;
 	private int Drop;
-	/*----------- Constrictors -----------*/	
-	
-	public Monstre(){
+	/*----------- Constrictors -----------*/
+
+	public Monstre() {
 	}
-	public Monstre(int id, String nom, String description, int niveau, Classe classe, int pvMax, int esquive,
+
+	public Monstre(Long id, String nom, String description, int niveau, Classe classe, int pvMax, int esquive,
 			int vitesse, int paMax, int pmMax, int attMagique, int attPhysique, int attDistance, int defMagique,
 			int defPhysique, int defDistance) {
-		
+
 		super(id, nom, description, niveau, classe, pvMax, esquive, vitesse, paMax, pmMax, attMagique, attPhysique,
 				attDistance, defMagique, defPhysique, defDistance);
 
 	}
-	
-
-
 
 	/*----------- Getters & Setters -----------*/
 
@@ -41,13 +38,8 @@ public class Monstre extends Vivant{
 		return Drop;
 	}
 
-
 	public void setDrop(int drop) {
 		Drop = drop;
 	}
-
-
-
-	
 
 }
